Refactor: Separate frontend and backend into organized file structure
- Extract Task model and TodoApp business logic to models/task.go
- Move HTML template to static/index.html with separate CSS and JS files
- Create handlers/api.go for all HTTP API endpoints
- Add GET /api/tasks endpoint for dynamic task loading
- Update main.go to serve static files and route to handlers
- Maintain all existing functionality while improving code organization
- Reduce main.go from 429 lines to 46 lines for better readability
